requirements.txt
# Core dependencies
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# HTTP client
aiohttp>=3.9.0

# Retry logic
tenacity>=8.2.0

# Web dashboard
flask>=3.0.0

# Configuration
python-dotenv>=1.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
